+++ CHANGELOG.md
@@ -33,6 +33,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   an upcoming release, at which point it will be sent to the underlying script,
   consistent with how npm usually behaves.
 
+- Scripts are no longer skipped as fresh if any `output` files were changed,
+  added, or removed since the previous run.
+
+- In order for a script to be skipped as fresh, it is now required to specify
+  the `output` files. Previously only input `files` were required.
+
 ## [0.5.0] - 2022-05-31
 
 ### Added